Refactor support for the `Heap` in wasm
Make it clear that the `Heap` isn't actually owned by the VM but is actually
taken by reference for various operations. Also fixup the types on `Opcode::opN`
to be `u16` instead of `i32`

Overall this should better prepare for transitioning the heap into wasm
entirely.

+++ packages/@glimmer/low-level/rust/src/ffi.rs
@@ -22,7 +22,7 @@ wasm_bindgen! {
         );
 
         pub fn low_level_vm_heap_get_addr(heap: &JsObject, at: u32) -> i32;
-        pub fn low_level_vm_heap_get_by_addr(heap: &JsObject, at: u32) -> u32;
+        pub fn low_level_vm_heap_get_by_addr(heap: &JsObject, at: u32) -> u16;
         pub fn debug_println(msg: &str);
     }
 }
